Beatmap adding OK
Beatmap adding to the database is now working as expected, but this process calculates 1 query for the map, then 144 queries for every mod combination permap, then for every of all of these mods, 3 accuracies values.

This is really slow, thinking about using an external db server starting with the app.